+++ MyFitnessJournal Watch App/Features/Main/Feature/MainFeatureAW+Action.swift
@@ -21,28 +21,35 @@ extension MainFeatureAW {
         
         // MARK: - View Actions
         
+        /// Handles user interactions or lifecycle events from the view layer.
         case view(View)
         
         enum View {
             
-            ///
+            /// Called when the view appears on screen.
             case viewDidAppear
             
+            /// Called when a workout option is selected from the UI.
             ///
+            /// - Parameter workout: The selected workout option.
             case selectedWorkoutOption(WorkoutOptionAW)
         }
         
         // MARK: - Actions
         
-        ///
+        /// Opens the training summary sheet.
         case openTrainingSheet
         
         // MARK: - Destination
         
+        /// Triggers navigation or logic based on the selected workout option.
         ///
+        /// - Parameter workout: The selected workout to show.
         case show(WorkoutOptionAW)
         
+        /// Handles navigation to a destination view or feature.
         ///
+        /// - Parameter action: The destination's scoped action.
         case destination(PresentationAction<Destination.Action>)
     }
