+++ MyFitnessJournal Watch App/Features/Main/Feature/MainFeatureAW+State.swift
@@ -17,7 +17,9 @@ extension MainFeatureAW {
         
         // MARK: - Properties
         
+        /// The list of workout options available to the user.
         ///
+        /// Default values include planned, mirroring, scheduled, and free workouts.
         var workoutTypes: [WorkoutOptionAW] = [.planned, .mirroring, .scheduled, .free]
         
         // MARK: - Destination
