+++ MyFitnessJournal Watch App/Features/Session/TrainingSession/Feature/TrainingSessionTabFeature.swift
@@ -88,7 +88,7 @@ struct TrainingSessionTabFeature {
                 }
                 
             case .metric(_):
-                return .none   
+                return .none
             }
         }
         Scope(state: \.controls, action: \.controls) {
@@ -110,16 +110,18 @@ extension TrainingSessionTabFeature {
         /// Action triggered when the user changes the selected tab.
         case tabChanged(WorkoutSessionScreenAW)
         
-        ///
+        /// Sets the selected workout activity type and triggers workout start.
         case setWorkoutActivityType(HKWorkoutActivityType)
         
+        /// Updates the state to reflect whether the workout session is running.
         ///
+        /// - Parameter isRunning: Boolean indicating the running state.
         case workoutSessionIsRunningChanged(Bool)
         
-        ///
+        /// Cancels active effects such as metrics or session monitoring streams.
         case cancelEffect
         
-        ///
+        /// Starts the workout session, including monitoring streams for session status and metrics.
         case workoutStart
         
         // MARK: - Actions
@@ -129,19 +131,19 @@ extension TrainingSessionTabFeature {
         /// Sub-actions for view-related events.
         enum View {
             
-            ///
+            /// Triggered when the view appears, used to initialize workout setup.
             case viewDidAppear
             
-            ///
+            /// User-triggered tab change to the workout screen.
             case changeTab
         }
         
         // MARK: - Child Actions
         
-        ///
+        /// Delegate action for training controls.
         case controls(TrainingControlsFeature.Action)
         
-        ///
+        /// Delegate action for workout metrics.
         case metric(TrainingMetricFeature.Action)
     }
     
@@ -155,21 +157,21 @@ extension TrainingSessionTabFeature {
         
         // MARK: - Properties
         
-        ///
+        /// The currently selected workout activity type, e.g., running, cycling.
         var selectedWorkout: HKWorkoutActivityType?
         
-        ///
+        /// Indicates whether the workout session is currently running.
         var workoutSessionIsRunning: Bool = false
         
-        ///
+        /// The currently selected tab in the workout session interface.
         var selectedTab: WorkoutSessionScreenAW = .workout
         
-        ///
+        /// The elapsed workout time used for tracking duration.
         var elapsedTime: TimeInterval = 0
         
         // MARK: - Child State
         
-        ///
+        /// State for managing the training control buttons and pause/resume status.
         var controls: TrainingControlsFeature.State {
             get {
                 .init(sessionIsRunning: workoutSessionIsRunning)
@@ -179,33 +181,18 @@ extension TrainingSessionTabFeature {
             }
         }
         
-        ///
+        /// State for displaying and managing real-time workout metrics.
         var metric = TrainingMetricFeature.State()
     }
     
 }
 
+/// Identifiers for cancellable effects used in the workout session feature.
+///
+/// These are used to manage long-running streams such as session state and metrics updates.
 private enum CancelId: Hashable {
+    /// Cancellation ID for the workout session running state stream.
     case cancelSessionRunning
+    /// Cancellation ID for the workout metrics stream.
     case cancelMetrics
 }
-
-
-// Client
-// SessionService
-//  var selectedWorkout: HKWorkoutActivityType?
-
-// MetricService
-//var workoutMetricsStream: AsyncStream<WorkoutMetrics> {
-//    workoutManager.workoutMetricsStream
-//}
-
-// contorls
-//var workoutSessionIsRunning: Bool {
-//    workoutManager.workoutSessionIsRunning
-//}
-
-//
-//func endWorkout() {
-//    workoutManager.endWorkout()
-//}
