+++ TASKS.md
@@ -380,4 +380,5 @@
     - Privacy - Health Share/Update Usage Description
     B: Authorization
     C: create WorkoutManager
-
+    - create training flow
+    - clean up folders
